Move deep_update into utils/misc.py
