--- wagtail/utils/utils.py
@@ -1,15 +1,0 @@
-import collections
-
-
-def deep_update(source, overrides):
-    """Update a nested dictionary or similar mapping.
-
-    Modify ``source`` in place.
-    """
-    for key, value in overrides.items():
-        if isinstance(value, collections.Mapping) and value:
-            returned = deep_update(source.get(key, {}), value)
-            source[key] = returned
-        else:
-            source[key] = overrides[key]
-    return source
